requirements.txt
annotated-types==0.7.0
anyio==4.8.0
arrow==1.3.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.8
codecarbon==2.8.2
contourpy==1.3.1
cryptography==44.0.0
cycler==0.12.1
fastapi==0.115.6
fief-client==0.20.0
filelock==3.16.1
fonttools==4.55.0
fsspec==2024.10.0
graphviz==0.20.3
h11==0.14.0
hiclass==4.12.0
httpcore==1.0.7
httpx==0.27.2
huggingface-hub==0.26.2
idna==3.10
Jinja2==3.1.4
joblib==1.4.2
jwcrypto==1.5.6
kiwisolver==1.4.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.2
mdurl==0.1.2
mpmath==1.3.0
networkx==3.4.2
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-ml-py==12.560.30
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
packaging==24.2
pandas==2.2.3
pillow==11.0.0
plotly==5.24.1
prometheus_client==0.21.1
prompt_toolkit==3.0.48
psutil==6.1.1
py-cpuinfo==9.0.0
pycparser==2.22
pydantic==2.10.5
pydantic_core==2.27.2
Pygments==2.19.1
pynvml==12.0.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
questionary==2.1.0
RapidFuzz==3.11.0
requests==2.32.3
rich==13.9.4
safetensors==0.4.5
scikit-learn==1.4.2
scipy==1.12.0
seaborn==0.13.2
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
starlette==0.41.3
sympy==1.13.1
tenacity==9.0.0
termcolor==2.3.0
threadpoolctl==3.5.0
timm==1.0.11
torch==2.5.1
torchvision==0.20.1
tqdm==4.67.0
triton==3.1.0
typer==0.15.1
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
wcwidth==0.2.13
yaspin==3.1.0
